+++ Project_2_Audio/lzwmap.h
@@ -7,13 +7,13 @@
 class LZWMap
 {
 public:
-    LZWMap(QVector<double> sequence);
+    LZWMap(QVector<float> sequence);
     int getSizeInBytes();
-    std::map<std::vector<double>, int> dictionary;
+    std::map<std::vector<float>, int> dictionary;
     std::vector<int> newCode;
 private:
-    int getCodeFromMap(std::vector<double> key, int * code);
-    void addToMap(std::vector<double> key, int value);
+    int getCodeFromMap(std::vector<float> key, int * code);
+    void addToMap(std::vector<float> key, int value);
 };
 
 #endif // LZWMAP_H
